sql-db/requirements.txt
mysql-connector==2.2.9
mysql-connector-python==9.0.0
python-dotenv==1.0.1
